refactor(web): migrate query-builder to feature folder
