+++ apps/web/src/tools/messages.ts
@@ -5,6 +5,7 @@ import { messages as jsonbQueryGenerator } from "./jsonb-query-generator/message
 import { messages as jwtDecoder } from "./jwt-decoder/messages";
 import { messages as mongoQueryGenerator } from "./mongo-query-generator/messages";
 import { messages as objectFlatten } from "./object-flatten/messages";
+import { messages as queryBuilder } from "./query-builder/messages";
 import { messages as typeConverter } from "./type-converter/messages";
 
 // As each tool is migrated, add its messages fragment here (i18n only, no component import).
@@ -15,4 +16,5 @@ export const toolMessages: LocaleMessages[] = [
   mongoQueryGenerator,
   jsonbQueryGenerator,
   jwtDecoder,
+  queryBuilder,
 ];
