+++ apps/web/src/tools/query-builder/index.ts
@@ -1,0 +1,5 @@
+import type { ToolModule } from "@/tools/types";
+
+import { QueryBuilder } from "./ui";
+
+export const tool: ToolModule = { id: "query-builder", Component: QueryBuilder };
